Refactored some old unneeded code
