+++ src/Battle.Logic/Map/FieldOfView.cs
@@ -1,9 +1,7 @@
 ﻿using Battle.Logic.Characters;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Numerics;
-using System.Text;
 
 namespace Battle.Logic.Map
 {
@@ -11,88 +9,24 @@ namespace Battle.Logic.Map
     {
         public static List<Vector3> GetFieldOfView(string[,] map, Vector3 location, int range)
         {
-            return MapCore.GetMapArea(map, location, range);
-            //int startingX = (int)location.X;
-            //int startingZ = (int)location.Z;
-            ////Use the range to find the borders in each primary direction from the starting location
-            //int minX = startingX - range;
-            //if (minX < 0)
-            //{
-            //    minX = 0;
-            //}
-            //int maxX = startingX + range;
-            //if (maxX > map.GetLength(0) - 1)
-            //{
-            //    maxX = map.GetLength(0) - 1;
-            //}
-            //int minZ = startingZ - range;
-            //if (minZ < 0)
-            //{
-            //    minZ = 0;
-            //}
-            //int maxZ = startingZ + range;
-            //if (maxZ > map.GetLength(0) - 1)
-            //{
-            //    maxZ = map.GetLength(0) - 1;
-            //}
+            return MapCore.GetMapArea(map, location, range, true);
+        }
 
-            ////Get a list of all border squares
-            //HashSet<Vector3> borderTiles = new();
-            ////Add the top and bottom rows
-            //for (int i = minX; i <= maxX; i++)
-            //{
-            //    borderTiles.Add(new(i, 0, minZ));
-            //    borderTiles.Add(new(i, 0, maxZ));
-            //}
-            ////Add the left and right sides
-            //for (int i = minZ; i < maxZ; i++)
-            //{
-            //    borderTiles.Add(new(minX, 0, i));
-            //    borderTiles.Add(new(maxX, 0, i));
-            //}
-
-            ////For each border tile, draw a line from the starting point to the border
-            //HashSet<Vector3> results = new();
-            //foreach (Vector3 borderItem in borderTiles)
-            //{
-            //    List<Vector3> singleLineCheck = FieldOfView.GetPointsOnLine(startingX, startingZ, (int)borderItem.X, (int)borderItem.Z).ToList<Vector3>();
-            //    if (singleLineCheck.Count > 0 &&
-            //        singleLineCheck[^1].X == startingX &&
-            //        singleLineCheck[^1].Z == startingZ) // note that ^1 is the same as singleLineCheck.Count - 1
-            //    {
-            //        //Reverse the list
-            //        singleLineCheck.Reverse();
-
-            //    }
-            //    double lineLength = MapCore.GetLengthOfLine(singleLineCheck[0], singleLineCheck[^1], 1);
-            //    double lineSegment = lineLength / singleLineCheck.Count;
-            //    double currentLength = 0;
-            //    for (int i = 0; i < singleLineCheck.Count; i++)
-            //    {
-            //        currentLength += lineSegment;
-            //        Vector3 fovItem = singleLineCheck[i];
-            //        //If we find an object, stop adding tiles
-            //        if (map[(int)fovItem.X, (int)fovItem.Z] == CoverType.FullCover)
-            //        {
-            //            break;
-            //        }
-            //        else if ((int)fovItem.X == startingX && (int)fovItem.Z == startingZ)
-            //        {
-            //            //Don't add this one, it's the origin/ where the character is looking from
-            //        }
-            //        else
-            //        {
-            //            results.Add(fovItem);
-            //        }
-            //        //We don't round, so this will extend the range a tiny part - but I think that is ok.
-            //        if (currentLength >= range)
-            //        {
-            //            break;
-            //        }
-            //    }
-            //}
-
-            //return results.ToList();
+        public static List<Character> GetCharactersInArea(List<Character> characters, string[,] map, Vector3 location, int radius)
+        {
+            List<Character> results = new();
+            List<Vector3> area = MapCore.GetMapArea(map, location, radius, false, true);
+            foreach (Character character in characters)
+            {
+                foreach (Vector3 item in area)
+                {
+                    if (character.Location == item)
+                    {
+                        results.Add(character);
+                    }
+                }
+            }
+            return results;
         }
 
         //http://ericw.ca/notes/bresenhams-line-algorithm-in-csharp.html
